refactor(types,wire): relocate orphan-rule-bound Value conversions after re-home (#1061)
Re-homing Value to reddb-io-types makes it foreign to reddb-server, so the
trait impls that pair Value with another foreign type can no longer live in
the server (orphan rule, E0117). Relocate each to the crate that owns its
non-Value side, byte-faithfully and with zero call-site edits:

- From<primitive> for Value (&str/String/i32/i64/f32/f64/bool): moved from
  reddb-server storage::unified::devx::conversions into the keystone crate
  (reddb_types::conversions) — Value's home.
- From<&Value>/From<Value> for WireValue and TryFrom<WireValue> for Value:
  moved from reddb-server wire::protocol into reddb-io-wire's legacy module
  (WireValue's home). reddb-wire gains a reddb-io-types dependency; the
  protocol helper fns (encode_value/decode_value) stay and resolve the impls
  via reddb_wire. listener.rs's WireValue::from(..) call-sites are untouched.

Added a focused round-trip test in legacy.rs pinning the WireValue<->Value
field mapping + the timestamp i64-overflow error path. cargo check --workspace
--locked green; cargo clippy --workspace -- -D warnings clean; fmt clean.
